Refactor Tool Serialization and Provider Integration: - Moved OpenAIFunctionListSerializer to provider.tools for better modularity. - Introduced SerialToolList for generic tool serialization. - Updated LocalAI and Ollama to use OpenAITool for serialization - Improved code structure and flexibility for future provider integration.
